Add coallesce
